refactor schema util functions
