+++ lib/document.js
@@ -7,6 +7,7 @@ var utils = require('./utils');
 var Schema = require('./schema');
 var Tree = require('./tree');
 var Type = require('./type');
+var cloneUtil = require('./clone-util');
 
 module.exports = Document;
 
@@ -439,13 +440,13 @@ Document.prototype.$_toObject = function (options, json) {
     (options && options.$_useSchemaOptions)) {
     if (json) {
       options = this.schema.options.toJSON ?
-        utils.clone(this.schema.options.toJSON) :
+        cloneUtil.clone(this.schema.options.toJSON) :
       {};
       options.json = true;
       options.$_useSchemaOptions = true;
     } else {
       options = this.schema.options.toObject ?
-        utils.clone(this.schema.options.toObject) :
+        cloneUtil.clone(this.schema.options.toObject) :
       {};
       options.json = false;
       options.$_useSchemaOptions = true;
@@ -462,7 +463,7 @@ Document.prototype.$_toObject = function (options, json) {
   // to save it from being overwritten by sub-transform functions
   var originalTransform = options.transform;
 
-  var ret = utils.clone(this.$_table, options) || {};
+  var ret = cloneUtil.clone(this.$_table, options) || {};
 
   if (options.virtuals) {
     applyGetters(this, ret, 'virtuals', options);
@@ -526,7 +527,7 @@ function applyGetters(self, json, type, options) {
     for (var ii = 0; ii < plen; ++ii) {
       part = parts[ii];
       if (ii === last) {
-        branch[part] = utils.clone(self.get(path), options);
+        branch[part] = cloneUtil.clone(self.get(path), options);
       } else {
         branch = branch[part] || (branch[part] = {});
       }
