+++ lib/index.js
@@ -221,7 +221,7 @@ Lounge.prototype.modelNames = function () {
       if (this.bucket) {
         return this.bucket[key].apply(this.bucket, arguments);
       }
-    }
+    };
   });
 
 /**
@@ -242,14 +242,14 @@ function propertyGetWrapper(key) {
     if (this.bucket) {
       return this.bucket[key];
     }
-  }
+  };
 }
 function propertySetWrapper(key) {
   return function (value) {
     if (this.bucket) {
       return this.bucket[key] = value;
     }
-  }
+  };
 }
 
 /**
@@ -299,4 +299,4 @@ Lounge.prototype.Document = Document;
 
 Lounge.prototype.Lounge = Lounge;
 
-module.exports = new Lounge;
+module.exports = new Lounge();
